+++ features/support/app_pages/admin/admin.rb
@@ -6,14 +6,18 @@ class AdminTool < IDmeBase
     super(FigNewton.admin.base_url)
   end
 
-  def login_in_new_window
-    self.create_new_window
-    self.use_last_browser_created
+  def login
     self.visit
     sleep 2
     authority = Authority.new
     sleep 2
     authority.auth_login
+  end
+
+  def login_in_new_window
+    self.create_new_window
+    self.use_last_browser_created
+    login
   end
 
   def logout_in_new_window