+++ features/support/app_pages/admin/member/user_quick_search.rb
@@ -1,0 +1,32 @@
+class UserQuickSearch < IDmeBase
+
+include Capybara::DSL
+include DataMagic
+
+  def initialize
+    super(FigNewton.admin.user_quick_search)
+  end
+
+  def populate_field(search_box:)
+    fill_in("query", :with => search_box)
+    click_button("Search")
+  end
+
+  def verify(populate: true, type: "none")
+    if populate
+      data = data_for(:admin_test_user)
+      case type
+      when "email"
+        populate_field(search_box: data.fetch("email"))
+        click_link(data.fetch("email"))
+      when "full_name"
+        populate_field(search_box: data.fetch("first_name") + " " + data.fetch("last_name"))
+        click_link(data.fetch("first_name") + " " + data.fetch("last_name"))
+      when "first_name"
+        populate_field(search_box: data.fetch("first_name"))
+        click_link(data.fetch("first_name"))
+      end
+    end
+    page.assert_text(data.fetch("first_name") + " " + data.fetch("last_name"))
+  end
+end