+++ features/support/app_pages/iva/military_dd214_vet.rb
@@ -6,7 +6,7 @@ class DD214 < IDmeBase
   include Capybara::DSL
   include ErrorMessages
 
-  def verify(affiliation: "Veteran", populate: true, type: "none", method:)
+  def verify(affiliation: "Veteran", populate: true, type: "none", method: "none")
     find("[data-option='dd214-request']").find(".verification-header").click
     populate_affiliation(affiliation)