+++ features/support/app_pages/shop/admin/shop_admin_dashboard_page.rb
@@ -1,0 +1,8 @@
+class ShopAdminDashboard < IDmeBase
+  include IDPBase
+
+  def initialize
+    super(FigNewton.shop_admin.base)
+  end
+
+end